# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Catálogo de Restaurantes",
        "description": "API completa para gestión de restaurantes, categorías de menú y productos",
        "contact": {
            "name": "Equipo de Desarrollo",
            "email": "admin@catalogo.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local de Desarrollo"
        }
    ],
    "paths": {
...
        "/api/restaurantes": {
            "get": {
                "tags": [
                    "Restaurantes"
                ],
                "summary": "Obtener lista de todos los restaurantes",
                "operationId": "df9c7769286ac2d81db4344712d71c82",
                "responses": {
                    "200": {
                        "description": "Lista de restaurantes obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Restaurante"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Restaurantes"
                ],
                "summary": "Crear un nuevo restaurante",
                "operationId": "7484e5c046190e6ab1c05d8daae3c6ac",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "usuario_admin_id",
                                    "nombre",
                                    "direccion",
                                    "telefono",
                                    "estado"
                                ],
                                "properties": {
                                    "usuario_admin_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "Mi Restaurante"
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "example": "Calle Principal 123"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "555-1234"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "enum": [
                                            "Abierto",
                                            "Cerrado"
                                        ],
                                        "example": "Abierto"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Restaurante creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Restaurante"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/restaurantes/{id}": {
            "get": {
                "tags": [
                    "Restaurantes"
                ],
                "summary": "Obtener un restaurante específico",
                "operationId": "ed6a5199dc793d61239a7f82cfb4c46a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del restaurante",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Restaurante encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Restaurante"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Restaurante no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Restaurantes"
                ],
                "summary": "Actualizar un restaurante existente",
                "operationId": "ec9a77a2b2bd5e106fb1bed350da0ec8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del restaurante",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "usuario_admin_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "Mi Restaurante Actualizado"
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "example": "Calle Principal 456"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "555-5678"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "enum": [
                                            "Abierto",
                                            "Cerrado"
                                        ],
                                        "example": "Cerrado"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Restaurante actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Restaurante"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Restaurante no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Restaurantes"
                ],
                "summary": "Eliminar un restaurante",
                "operationId": "5b8870f200f346e843e4ffc274e7aef5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del restaurante",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Restaurante eliminado exitosamente"
                    },
                    "404": {
                        "description": "Restaurante no encontrado"
                    }
                }
            }
        },
        "/api/restaurantes/usuario/{usuarioAdminId}": {
            "get": {
                "tags": [
                    "Restaurantes"
                ],
                "summary": "Obtener restaurantes por usuario administrador",
                "operationId": "eee519aed0add4b0b491a30b582f884c",
                "parameters": [
                    {
                        "name": "usuarioAdminId",
                        "in": "path",
                        "description": "ID del usuario administrador",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de restaurantes obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Restaurante"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Restaurante": {
                "title": "Restaurante",
                "description": "Modelo de Restaurante",
                "required": [
                    "usuario_admin_id",
                    "nombre",
                    "direccion",
                    "telefono",
                    "estado"
                ],
                "properties": {
                    "id": {
                        "description": "ID del restaurante",
                        "type": "integer",
                        "format": "int64"
                    },
                    "usuario_admin_id": {
                        "description": "ID del usuario administrador",
                        "type": "integer"
                    },
                    "nombre": {
                        "description": "Nombre del restaurante",
                        "type": "string"
                    },
                    "direccion": {
                        "description": "Dirección del restaurante",
                        "type": "string"
                    },
                    "telefono": {
                        "description": "Teléfono del restaurante",
                        "type": "string"
                    },
                    "estado": {
                        "description": "Estado del restaurante",
                        "type": "string",
                        "enum": [
                            "Abierto",
                            "Cerrado"
                        ]
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha de actualización",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}